+++ Changelog.md
@@ -5,11 +5,16 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+
+
+## [0.1.2]-2022-01-15
+
 ### Add
 
-- Image format setting.
-
-
+- Overwrite file option
+- Image write type (main thread / Async)
+- Filename serialized option (-0/-1/-2 ...)
+- Image format option
 
 ## [0.1.1] - 2022-01-12
 ### Changed